resolve conflicts and move some files to be in shared package.
