+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/component/AddOrEditTaskBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.tasks
+package com.amsterdam.cutetudee.presentation.screens.component
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.background
@@ -37,12 +37,13 @@ import com.amsterdam.cutetudee.presentation.component.ReadOnlyCustomTextField
 import com.amsterdam.cutetudee.presentation.component.SelectedBadgedCategory
 import com.amsterdam.cutetudee.presentation.component.chip.priority.PriorityChip
 import com.amsterdam.cutetudee.presentation.component.chip.priority.PriorityUi
-import com.amsterdam.cutetudee.presentation.screens.tasks.AddEditTaskUiState.CategoryItemUiState
+import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskInteractionListener
+import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskUiState
+import com.amsterdam.cutetudee.presentation.screens.common.AddEditTaskUiState.CategoryItemUiState
 import com.amsterdam.cutetudee.presentation.theme.AppTheme
 import com.amsterdam.cutetudee.presentation.theme.CuteTudeeTheme
 import com.amsterdam.cutetudee.presentation.utils.ThemeAndLocalePreviews
 import com.amsterdam.cutetudee.presentation.utils.dropShadow
-import kotlinx.datetime.LocalDate
 import kotlin.uuid.ExperimentalUuidApi
 
 @OptIn(ExperimentalUuidApi::class, ExperimentalMaterial3Api::class)
